requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
tensorflow-cpu==2.16.1
numpy==1.26.4
psycopg2-binary==2.9.9
boto3==1.34.133
joblib==1.4.2
google-generativeai
scikit-learn
pandas
